#13 Refactor code, add javadoc
