+++ src/test/java/com/exasol/versionnumberprovider/LocalExasolVersionNumberProviderFactory.java
@@ -5,16 +5,16 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 
 import com.exasol.versionnumberprovider.dockerhub.Deserializer;
-import com.exasol.versionnumberprovider.dockerhub.TagList;
+import com.exasol.versionnumberprovider.dockerhub.TagListPage;
 
 public class LocalExasolVersionNumberProviderFactory {
 
     public ExasolVersionNumberProvider getLocalDockerHubResponseDump(final Path path) throws IOException {
-        final TagList tagList = readTagList(path);
+        final TagListPage tagList = readTagList(path);
         return new ExasolVersionNumberProvider(tagList.getTags());
     }
 
-    private TagList readTagList(final Path path) throws IOException {
+    private TagListPage readTagList(final Path path) throws IOException {
         return new Deserializer().deserialize(Files.readString(path));
     }
 }
